Create Playlist Maker
A database was created of songs and artists. In this first step, the title of all the songs by the artist named 'Queen' were selected.
A 'Pop' playlist was created. In preparation, selected the name of all of the artists from the 'Pop' genre.
Finished by creating the 'Pop' playlist, added another query that will select the title of all the songs from the 'Pop' artists. Used IN on a nested subquery that's based on the previous query.