Create Gradebook
Created a database to track student grades, with their name, number grade, and what percent of activities they've completed. In this first step, select all of the rows, and display the name, number_grade, and percent_completed, which can compute by multiplying and rounding the fraction_completed column.

The goal is to show how many students have earned which letter_grade. Output the letter_grade by using CASE with the number_grade column, outputting 'A' for grades > 90, 'B' for grades > 80, 'C' for grades > 70, and 'F' otherwise. Used COUNT with GROUP BY to show the number of students with each of those grades.