Create Customer's orders
Came up with a query that lists the name and email of every customer followed by the item and price of orders they've made. Used a LEFT OUTER JOIN so that a customer is listed even if they've made no orders, and didn't add any ORDER BY.
Created another query that will result in one row per each customer, with their name, email, and total amount of money they've spent on orders. Sort the rows according to the total money spent, from the most spent to the least spent.